Remove URITemplate, fix failing test
